+++ Runtime/Jobs/UpdateJobOptionsAttribute.cs
@@ -2,7 +2,7 @@ using System;
 
 namespace Gilzoide.UpdateManager.Jobs
 {
-    [AttributeUsage(AttributeTargets.Struct), Obsolete("Use JobBatchSizeAttribute and ReadOnlyTransformsAttribute instead")]
+    [AttributeUsage(AttributeTargets.Struct), Obsolete("Use JobBatchSizeAttribute and ReadOnlyTransformAccessAttribute instead")]
     public class UpdateJobOptionsAttribute : Attribute
     {
         public int BatchSize { get; set; }